Add files via upload
To solve these challenges students form teams and work together for 2 to 6 months. These teams are formed in a way that they have students from different levels of education (from high school to Ph.D. level) and with different backgrounds (from tourism to nuclear physics). Throughout the course, they will learn as a team problem solving and product development skills and while applying those skills to develop a functioning prototype. To achieve this goal they come up with hundreds of ideas, which they will test, and through experimentation and cheap and fast concept failure, they narrow the solution down to one, final prototype. In the past three years over 300 students in Finland only went through the Epic Challenge educational program.

The novelty in the EPIC-project is in the way of how hundreds of students and researchers work on highly complex innovation challenges as part of their studies and how the SMEs in the region are integrated into the solution development process.